add line numbers and clean up code a little bit
