+++ utils.py
@@ -1,87 +1,32 @@
-import tkinter as tk
-from tkinter import TclError
+import uuid
+import requests
 
-# class TextLineNumbers(tk.Canvas):
-#     def __init__(self, *args, **kwargs):
-#         tk.Canvas.__init__(self, *args, **kwargs)
-#         self.textwidget = None
+def generate_unique_name():
+    unique_name = str(uuid.uuid4()) + ".py"
+    return unique_name
 
-#     def attach(self, text_widget):
-#         self.textwidget = text_widget
+def send_post_request(payload):
+    url = "https://baxin-simulator.hf.space/protocol"
+    protocol_name = generate_unique_name()
+    data = {"name": protocol_name, "content": payload}
+    headers = {"Content-Type": "application/json"}
 
-#     def redraw(self, *args):
-#         self.delete("all")
-#         i = self.textwidget.index("@0,0")
-#         while True:
-#             dline= self.textwidget.dlineinfo(i)
-#             if dline is None:
-#                 break
-#             y = dline[1]
-#             linenum = str(i).split(".")[0]
-#             self.create_text(2,y,anchor="nw", text=linenum)
-#             i = self.textwidget.index("%s+1line" % i)
+    response = requests.post(url, json=data, headers=headers)
 
-class CustomText(tk.Text):
-    def __init__(self, *args, **kwargs):
-        tk.Text.__init__(self, *args, **kwargs)
-        self.vsb = tk.Scrollbar(self, orient="vertical", command=self.text.yview)
-        self.configure(yscrollcommand=self.vsb.set)
-        self.linenumbers = TextLineNumbers(self, width=30)
-        self.linenumbers.attach(self)
-        self.vsb.pack(side="right", fill="y")
-        self.linenumbers.pack(side="left", fill="y")
-        self.pack(side="right", fill="both", expand=True)
-        self.bind("<<Change>>", self._on_change)
-        self.bind("<Configure>", self._on_change)
+    if response.status_code != 200:
+        print("Error: " + response.text)
+        return "Error: " + response.text
 
-    def _on_change(self, event):
-        self.linenumbers.redraw()
+    response_data = response.json()
+    if "error_message" in response_data:
+        print("Error in response:", response_data["error_message"])
+        return response_data["error_message"]
+    elif "protocol_name" in response_data:
+        return response_data["run_status"]
+    else:
+        print("Unexpected response:", response_data)
+        return "Unexpected response"
 
-
-# class CustomText(tk.Text):
-#     def __init__(self, *args, **kwargs):
-#         tk.Text.__init__(self, *args, **kwargs)
-#         # create a proxy for the underlying widget
-#         self._orig = self._w + "_orig"
-#         self.tk.call("rename", self._w, self._orig)
-#         self.tk.createcommand(self._w, self._proxy)
-#     def _proxy(self, *args):
-#         # let the actual widget perform the requested action
-#         cmd = (self._orig,) + args
-#         try:
-#             result = self.tk.call(cmd)
-#         except TclError:
-#             pass
-#         # generate an event if something was added or deleted,
-#         # or the cursor position changed
-#         if (args[0] in ("insert", "replace", "delete") or
-#             args[0:3] == ("mark", "set", "insert") or
-#             args[0:2] == ("xview", "moveto") or
-#             args[0:2] == ("xview", "scroll") or
-#             args[0:2] == ("yview", "moveto") or
-#             args[0:2] == ("yview", "scroll")
-#         ):
-#             self.event_generate("<<Change>>", when="tail")
-#         # return what the actual widget returned
-#         return result
-
-
-class TextLineNumbers(tk.Canvas):
-    def __init__(self, *args, **kwargs):
-        tk.Canvas.__init__(self, *args, **kwargs)
-        self.textwidget = None
-    def attach(self, text_widget):
-        self.textwidget = text_widget
-
-    def redraw(self, *args):
-        '''redraw line numbers'''
-        self.delete("all")
-        i = self.textwidget.index("@0,0")
-        while True :
-            dline= self.textwidget.dlineinfo(i)
-            if dline is None: break
-            y = dline[1]
-            linenum = str(i).split(".")[0]
-            self.create_text(2,y,anchor="nw", text=linenum)
-            i = self.textwidget.index("%s+1line" % i)
-
+def send_code(text):
+   response = send_post_request(text)
+   return response
